build bugs?
Noticed that the NodeValue.h file was using functions from StringMap but the header wasn't included.

The BundleSaver and LLVMIRGen looks like you copied and pasted the same code you wanted to apply to FB and to non-FB computers.

The last issue was with llvm::make_unique. llvm::make_unique wants to transition to std::make_unique (https://reviews.llvm.org/D66259) so maybe this was why is wasn't building.

In the end it built with some warnings. I noticed there were a few other potential issues, but I have yet to run the whole program so I don't know if these issues actually matter.